+++ src/Relude/Functor/Reexport.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE CPP  #-}
 {-# LANGUAGE Safe #-}
 
 {- |
@@ -15,9 +16,18 @@ module Relude.Functor.Reexport
        , module Data.Functor
        , module Data.Functor.Compose
        , module Data.Functor.Identity
+#if MIN_VERSION_base(4,12,0)
+       , module Data.Functor.Contravariant
+#endif
        ) where
 
 import Data.Bifunctor (Bifunctor (..))
 import Data.Functor (Functor (..), void, ($>), (<$>))
 import Data.Functor.Compose (Compose (..))
 import Data.Functor.Identity (Identity (..))
+
+#if MIN_VERSION_base(4,12,0)
+import Data.Functor.Contravariant (Comparison (..), Contravariant (..), Equivalence (..), Op (..),
+                                   Predicate (..), comparisonEquivalence, defaultComparison,
+                                   defaultEquivalence, phantom, ($<), (>$$<), (>$<))
+#endif